refactored schemas and indexing, update api for survey pipeline
