+++ doc/documentation/autocomplete/index.rst
@@ -3,8 +3,3 @@ Autocomplete
 
 .. automodule:: autocomplete.autocompleter
    :members:
-
-.. toctree::
-   :hidden:
-
-   matching
