--- doc/documentation/autocomplete/matching.rst
@@ -1,6 +1,0 @@
-Matching
-########
-
-
-.. automodule:: database.service.matching
-   :members:
