--- doc/documentation/database/base.rst
@@ -1,5 +1,0 @@
-Base
-####
-
-.. automodule:: database.service.base
-   :members:
