--- doc/documentation/database/document.rst
@@ -1,5 +1,0 @@
-Document
-########
-
-.. automodule:: database.service.document
-   :members:
