--- doc/documentation/database/matching.rst
@@ -1,5 +1,0 @@
-Matching
-########
-
-.. automodule:: database.service.matching
-   :members:
