--- doc/documentation/database/question.rst
@@ -1,5 +1,0 @@
-Question
-########
-
-.. automodule:: database.service.question
-   :members:
