+++ doc/documentation/database/services/base.rst
@@ -1,0 +1,5 @@
+Base
+####
+
+.. autoclass:: database.service.base.BaseService
+   :members:
