+++ doc/documentation/database/services/document.rst
@@ -1,0 +1,5 @@
+Document
+########
+
+.. autoclass:: database.service.document.DocumentService
+   :members:
