+++ doc/documentation/database/services/embedding.rst
@@ -1,0 +1,5 @@
+Embedding
+#########
+
+.. autoclass:: database.service.base.EmbeddingService
+   :members:
