+++ doc/documentation/database/services/index.rst
@@ -1,0 +1,12 @@
+Services
+========
+
+.. toctree::
+   :hidden:
+
+   base
+   embedding
+   matching
+   document
+   question
+   source
