+++ doc/documentation/database/services/matching.rst
@@ -1,0 +1,5 @@
+Matching
+########
+
+.. autoclass:: database.service.matching.MatchingService
+   :members:
