+++ doc/documentation/database/services/question.rst
@@ -1,0 +1,5 @@
+Question
+########
+
+.. autoclass:: database.service.question.QuestionService
+   :members:
