--- doc/documentation/database/source.rst
@@ -1,5 +1,0 @@
-Source
-######
-
-.. automodule:: database.service.source
-   :members:
