+++ doc/documentation/database/tables.rst
@@ -1,0 +1,19 @@
+Tables
+======
+
+As stated previously, the database structure is defined in the :mod:`database.models` module. Currently, it is composed of the following tables.
+
+Document
+--------
+
+.. sqla-model:: database.models.Document
+
+Question
+--------
+
+.. sqla-model:: database.models.Question
+
+Source
+------
+
+.. sqla-model:: database.models.Source
