+++ doc/documentation/indexing/index.rst
@@ -1,13 +1,14 @@
-IndexingPipeline
-################
+Indexing
+########
 
-.. automodule:: indexing_api
-   :members:
+Create a pipeline to index data from a source to the database, using a parser and a scraper.
+
+.. autoclass:: indexing.base.BaseIndexer
+    :members:
 
 .. toctree::
    :hidden:
 
-   crawler
-   scraper
    parser
-   chunker
+   scraper
+   pipelines/index
