+++ doc/documentation/indexing/pipelines/ahv.rst
@@ -1,0 +1,7 @@
+AHV/IV pipeline
+===============
+
+Index the website https://www.ahv-iv.ch/ for RAG data.
+
+.. automodule:: indexing.pipelines.ahv
+   :members:
