+++ doc/documentation/indexing/pipelines/bsv.rst
@@ -1,0 +1,7 @@
+BSV pipeline
+============
+
+Index the website https://www.bsv.admin.ch/ for Autocomplete data.
+
+.. automodule:: indexing.pipelines.bsv
+   :members:
