+++ doc/documentation/indexing/pipelines/eak.rst
@@ -1,0 +1,7 @@
+EAK pipeline
+============
+
+Index the website https://www.eak.admin.ch/ for RAG data.
+
+.. automodule:: indexing.pipelines.admin
+   :members:
