+++ doc/documentation/indexing/pipelines/index.rst
@@ -1,0 +1,9 @@
+Indexing pipelines
+==================
+
+.. toctree::
+   :hidden:
+
+   ahv
+   eak
+   bsv
