+++ doc/documentation/indexing/scraper.rst
@@ -1,5 +1,8 @@
 Scraper
 #######
 
+.. autoclass:: indexing.base.BaseScraper
+    :members:
+
 .. autoclass:: indexing.scraper.Scraper
     :members:
