+++ doc/documentation/schemas/document.rst
@@ -1,0 +1,5 @@
+Document
+========
+
+.. automodule:: schemas.document
+    :members:
