+++ doc/documentation/schemas/question.rst
@@ -1,0 +1,5 @@
+Question
+========
+
+.. automodule:: schemas.question
+    :members:
