+++ doc/documentation/schemas/source.rst
@@ -1,0 +1,5 @@
+Source
+======
+
+.. automodule:: schemas.source
+    :members:
