+++ src/copilot/app/database/service/base.py
@@ -1,11 +1,12 @@
 from sqlalchemy import select
 from sqlalchemy.orm import Session
-from ..models import Base
+from ..models import Base, EmbeddedMixin
+from pydantic import BaseModel
 
 from utils.embedding import get_embedding
 
 from abc import ABCMeta, abstractmethod
-from typing import Type, Union, Tuple
+from typing import Type
 
 import logging
 logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(name)s - %(levelname)s - %(message)s')
@@ -13,6 +14,14 @@ logger = logging.getLogger(__name__)
 
 
 class BaseService(metaclass=ABCMeta):
+    """
+    Base class for database services
+
+    Parameters
+    ----------
+    model
+        Database model
+    """
     def __init__(self, model=Type[Base]):
         self.model = model
 
@@ -21,12 +30,42 @@ class BaseService(metaclass=ABCMeta):
         db.add(db_obj)
         return db_obj
 
-    def create(self, db: Session, obj_in):
+    def create(self, db: Session, obj_in: BaseModel):
+        """
+        Create a new object in the database
+
+        Parameters
+        ----------
+        db : Session
+            Database session
+        obj_in
+            Pydantic schema
+
+        Returns
+        -------
+        Base
+            Database object
+        """
         db_obj = self._create(db, obj_in)
         db.commit()
         return db_obj
 
-    def create_all(self, db: Session, objs_in):
+    def create_all(self, db: Session, objs_in: BaseModel):
+        """
+        Create multiple objects in the database
+
+        Parameters
+        ----------
+        db: Session
+            Database session
+        objs_in
+            Pydantic schema with list of objects to create
+
+        Returns
+        -------
+        list[Base]
+            List of database objects
+        """
         db_objs = []
         for obj_in in objs_in.objects:
             db_objs.append(self._create(db, obj_in))
@@ -34,26 +73,50 @@ class BaseService(metaclass=ABCMeta):
         return db_objs
 
     def get(self, db: Session, id_: int):
+        """
+        Get object by id
+
+        Parameters
+        ----------
+        db: Session
+            Database session
+        id_: int
+            Object id
+
+        Returns
+        -------
+        Base
+            Database object
+        """
         stmt = select(self.model).filter(self.model.id == id_)
         return db.execute(stmt).scalar_one()
 
-    def _update(self, db: Session, db_obj, obj_in):
-        for field, value in obj_in.model_dump(exclude_unset=True).items():
-            setattr(db_obj, field, value)
-        return db_obj
+    def update(self, db: Session, db_obj: Base, obj_in: BaseModel):
+        """
+        Update db object
 
-    def update(self, db: Session, db_obj, obj_in):
+        Parameters
+        ----------
+        db : Session
+            Database session
+        db_obj : Base
+            Database object
+        obj_in : BaseModel
+            Pydantic schema
+
+        Returns
+        -------
+        Base
+            Updated database object
+        """
         db_obj = self._update(db, db_obj, obj_in)
         db.commit()
         return db_obj
 
-    def update_all(self, db: Session, objs_in):
-        db_objs = []
-        for obj_in in objs_in:
-            db_obj = self.get(db, obj_in.id)
-            db_objs.append(self.update(db, db_obj, obj_in))
-        db.commit()
-        return db_objs
+    def _update(self, db: Session, db_obj: Base, obj_in: BaseModel):
+        for field, value in obj_in.model_dump(exclude_unset=True).items():
+            setattr(db_obj, field, value)
+        return db_obj
 
     def _delete(self, db: Session, id_: int):
         db_obj = self.get(db, id_)
@@ -61,11 +124,41 @@ class BaseService(metaclass=ABCMeta):
         return db_obj
 
     def delete(self, db: Session, id_: int):
+        """
+        Delete object by id
+
+        Parameters
+        ----------
+        db : Session
+            Database session
+        id_ : int
+            Object id
+
+        Returns
+        -------
+        Base
+            Deleted database object
+        """
         db_obj = self._delete(db, id_)
         db.commit()
         return db_obj
 
     def delete_all(self, db: Session, ids: list[int]):
+        """
+        Delete multiple objects by id
+
+        Parameters
+        ----------
+        db : Session
+            Database session
+        ids : list[int]
+            List of object ids
+
+        Returns
+        -------
+        list[Base]
+            List of deleted database objects
+        """
         db_objs = []
         for id_ in ids:
             db_objs.append(self._delete(db, id_))
@@ -74,20 +167,55 @@ class BaseService(metaclass=ABCMeta):
 
 
 class EmbeddingService(BaseService):
+    """
+    Base class for embedding services
+    """
 
-    def _embedding(self,db_obj, text: str):
+    def _embedding(self, db_obj: EmbeddedMixin, text: str):
         db_obj.embedding = get_embedding(text)
         return db_obj
 
-    def _embed(self, db_obj):
+    def _embed(self, db_obj: EmbeddedMixin):
         return self._embedding(db_obj, db_obj.text)
 
-    def embed_one(self, db: Session, db_obj):
+    def embed_one(self, db: Session, db_obj: EmbeddedMixin):
+        """
+        Embed a single object in the database
+
+        Parameters
+        ----------
+        db : Session
+            Database session
+        db_obj : EmbeddedMixin
+            Database object
+
+        Returns
+        -------
+        EmbeddedMixin
+            Embedded database object
+        """
         db_obj = self._embed(db_obj)
         db.commit()
         return db_obj
 
-    def embed_many(self, db: Session, embed_empty_only=True, k: int = 0):
+    def embed_many(self, db: Session, embed_empty_only: bool = True, k: int = 0):
+        """
+        Embed multiple objects in the database
+
+        Parameters
+        ----------
+        db : Session
+            Database session
+        embed_empty_only : bool, optional
+            Whether to embed only objects with empty embeddings (set to None) or all objects without distinction, default to True
+        k : int, optional
+            Number of objects to embed, with 0 meaning all objects, default to 0
+
+        Returns
+        -------
+        list[EmbeddedMixin]
+            List of embedded database objects
+        """
         stmt = select(self.model)
         if embed_empty_only:
             stmt = stmt.filter(self.model.embedding.is_(None))
@@ -104,15 +232,32 @@ class EmbeddingService(BaseService):
         return db_objs
 
     @abstractmethod
-    def _create(self, db: Session, obj_in, embed=False):
+    def _create(self, db: Session, obj_in: BaseModel, embed=False):
         pass
 
-    def create(self, db: Session, obj_in, embed=False):
+    def create(self, db: Session, obj_in: BaseModel, embed=False):
         db_obj = self._create(db, obj_in, embed=embed)
         db.commit()
         return db_obj
 
-    def create_all(self, db: Session, objs_in, embed=False):
+    def create_all(self, db: Session, objs_in: BaseModel, embed=False):
+        """
+        Create multiple objects in the database
+
+        Parameters
+        ----------
+        db : Session
+            Database session
+        objs_in : BaseModel
+            Pydantic schema with list of objects to create
+        embed : bool, optional
+            Whether to embed the objects to create, default to False
+
+        Returns
+        -------
+        list[Base]
+            List of database objects
+        """
         db_objs = []
         for obj_in in objs_in.objects:
             db_objs.append(self._create(db, obj_in, embed=embed))
@@ -120,6 +265,21 @@ class EmbeddingService(BaseService):
         return db_objs
 
     def get_by_text(self, db: Session, text: str):
+        """
+        Get object by text
+
+        Parameters
+        ----------
+        db : Session
+            Database session
+        text : str
+            Object text
+
+        Returns
+        -------
+        Base
+            Database object
+        """
         return db.query(self.model).filter(self.model.text == text).one_or_none()
 
     def _update(self, db: Session, db_obj, obj_in, embed=False):
@@ -142,10 +302,46 @@ class EmbeddingService(BaseService):
         logger.info(f'Excluded fields: {exclude}')
         return exclude
 
-    def update(self, db: Session, db_obj, obj_in, embed=False):
+    def update(self, db: Session, db_obj, obj_in: BaseModel, embed=False):
+        """
+        Update db object
+
+        Parameters
+        ----------
+        db : Session
+            Database session
+        db_obj
+            Database object
+        obj_in : BaseModel
+            Pydantic schema
+        embed : bool, optional
+            Whether to embed the object to update, default to False
+
+        Returns
+        -------
+        Base
+            Updated database object
+        """
         return self._update(db, db_obj, obj_in, embed=embed)
 
-    def upsert(self, db: Session, obj_in, embed=False):
+    def upsert(self, db: Session, obj_in: BaseModel, embed=False):
+        """
+        Upsert object in the database
+
+        Parameters
+        ----------
+        db : Session
+            Database session
+        obj_in : BaseModel
+            Pydantic schema
+        embed : bool, optional
+            Whether to embed the object to upsert, default to False
+
+        Returns
+        -------
+        Base
+            Upserted database object
+        """
         db_obj = self.get_by_text(db, obj_in.text)
         if db_obj:
             db_obj = self._update(db, db_obj, obj_in, embed=embed)
