--- src/copilot/app/indexing/models.py
@@ -1,10 +1,0 @@
-from pydantic import BaseModel
-from typing import Optional
-
-
-class FaqItem(BaseModel):
-    id: Optional[int] = None
-    question: str
-    answer: str
-    url: str
-    language: str
